+++ CHANGELOG.md
@@ -144,6 +144,7 @@
 121. Fixed Issue with ValueEdited not being set when Auto-Converter runs.
 122. Added UnityEngine.Il2CppImageConversionManager to Managed folder to help with Image Conversion on Il2Cpp Games.
 123. Added Better Exception Handling for MelonLoader Core Initialization.
+124. Added Error Log File Output to Installer.
 
 ---